refactor: change monad to interface
