+++ src/types/Monad.js
@@ -1,0 +1,5 @@
+// @flow
+
+export interface Monad<T> {
+  bind<U>(transform: (value: T) => Monad<U>): Monad<U>;
+}
